Add 'usb' option back to sign_official_build.sh script
We still need a way to re-sign non-installer images so that they can be
booted directly from USB.

BUG=chromium-os:8679
TEST=manual, from within the build chroot

Obtain a chromiumos_base_image from buildbot or your own build. Ensure that
it's signed with the dev-keys (it should be).

Modify it somehow. For example:

  (cros-chroot)$  cd src/platform/vboot_reference/scripts/image_signing
  (cros-chroot)$  ./set_chronos_password.sh chromiumos_base_image.bin mypassword

Now resign the image:

  (cros-chroot)$  cd src/platform/vboot_reference/scripts/image_signing
  (cros-chroot)$  ./sign_official_build.sh usb chromiumos_base_image.bin \
                    /usr/share/vboot/devkeys usb_image.bin

Then copy the usb_image to a USB stick:

  sudo dd if=usb_image of=/dev/WHATEVER

The resulting USB stick should boot in recovery mode, and assuming you
changed the password as shown above, should let you use that password to get
a shell.

Change-Id: I3aaa2b8787c52940249fd15007e075de7e017d78

Review URL: http://codereview.chromium.org/4424003