Remove -isystem from CFLAGS for firmware build
The -isystem and the rest of the CFLAGS for firmware builds is copied
from U-Boot, where U-Boot generates it on the fly, as a temporary
solution before we figure out how make the CFLAGS right.

Given that, the hard-coded -isystem is both incorrect (since tool chain
is upgraded to a new version) and unnecessary.  It is unnecessary
because firmware lib is carefully written that the lib does not (and
probably should not) depend on any system header.

Even if in the future a system header is added to the firmware lib,
because firmware build sets -nostdinc to CFLAGS, the compiler will
safely report missing header instead of silently using the standard
system header.

So this commit removes the -isystem.

BUG=chromium-os:16808
TEST=Make sure non-firmware build still works by running
     `emerge-{tegra2_seaboard,x86-alex} vboot_reference`
TEST=Run firmware build successfully
     `emerge-{tegra2_seaboard,x86-alex} vboot_reference-firmware`
TEST=Add #include<stdarg.h> to any header in firmware/include/ and run
     firmware build again and observe build fail on missing stdarg.h

Change-Id: I8291390f21a975446993640d7a92a3eed4750e32
Reviewed-on: http://gerrit.chromium.org/gerrit/10072
Tested-by: Che-Liang Chiou <clchiou@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Randall Spangler <rspangler@chromium.org>