Add Stumpy MP key info to vbutil_what_keys
BUG=chrome-os-partner:7878
TEST=manual

Run vbutil_what_keys specifying either a BIOS or disk image signed with the
Stumpy MP keys. It should identify it as such. For example:

  vbutil_what_keys  chromeos_1675.0.0_stumpy_recovery_dev-channel_mp.bin

or

  vbutil_what_keys bios.bin

The output should contain the strings "Stumpy MP" somewhere, if the image or
BIOS is signed with the Stumpy MP keys.

Change-Id: I575b7358ced4234c918eff40cdeb17fe06ab331c
Reviewed-on: https://gerrit.chromium.org/gerrit/15271
Commit-Ready: Bill Richardson <wfrichar@chromium.org>
Tested-by: Bill Richardson <wfrichar@chromium.org>
Reviewed-by: Randall Spangler <rspangler@chromium.org>