Fix incorrect CFLAGS when compiling for host
Both HOSTLIB and FWLIB depend on FWLIB_OBJS, and it's the OBJS which
need to override their CFLAGS.  The way it was, depending on the order
of evaluation of HOSTLIB and FWLIB, the OBJS would get compiled with
different flags.

So 'make' worked, because it depended on fwlib before hostlib.  But
'make runtests' failed, because it depended on hostlib but not fwlib.

BUG=none
BRANCH=none
TEST=make runtests

Change-Id: Id4e4cb4f7a58c7ab8d9fd00d91c4ab8f4518d4bc
Signed-off-by: Randall Spangler <rspangler@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/42014
Reviewed-by: Bill Richardson <wfrichar@chromium.org>