Avoid exit code overflow for tpmc.
In case of a TPM error, tpmc returns the TPM error code, which can
be greater than 255.  In that case the error code is truncated.
Some error codes, such as TPM_E_RETRY, end with a zero byte,
resulting in a successful exit code.  This is despicable.

BUG=chromium:234357
TEST=tested with exit codes < 255.  Too hard to generate the others.
BRANCH=none

Change-Id: I891a5c0659c06aac778449e2a0a935c5f82ccdb8
Reviewed-on: https://chromium-review.googlesource.com/66885
Reviewed-by: Luigi Semenzato <semenzato@chromium.org>
Commit-Queue: Luigi Semenzato <semenzato@chromium.org>
Tested-by: Luigi Semenzato <semenzato@chromium.org>