AMD Steppe Eagle: Add Platform Security Processor Binaries
Steppe Eagle (00730F01) contains the Avalon southbridge and
a Platform Security Processor (PSP). Supporting the PSP requires
specific binaries to be included in the rom. These binaries are
being added to the AMD directory as separate pieces but will be
swapped out for a combined binary in the future. The fletcher
utility is used to sign PSP binaries.

Change-Id: If6325d5f9ecec141317436a602c4cc1349a3f13f
Signed-off-by: Bruce Griffith <Bruce.Griffith@se-eng.com>