vboot: Add system-level test for LoadKernel()
This creates a disk image and verifies a kernel can be loaded from it.
It is roughly analogous to vb2_firmware_tests.sh, but at the kernel
step instead of the firmware step.

This will get more interesting in the near future, with the upcoming
addition of a streaming API to read the kernel.

BUG=chromium:408265
BRANCH=none
TEST=make runtests

Change-Id: Icc9e6d0e318c4bd38fc9ab1ad704da99232822e1
Signed-off-by: Randall Spangler <rspangler@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/214508
Reviewed-by: Daisuke Nojiri <dnojiri@chromium.org>