futility: Be more accepting of how it's invoked
Now that futility is pretty much working as intended, we don't
have to be quite so picky in the way it's being invoked. Up until
now, it's only worked when invoked as "futility" or as one of the
built-in commands, such as "dump_fmap".

This change removes those restrictions. You can invoke futility
under any name you wish. If it recognizes the name as a built-in
command, great. Otherwise it will require a valid command as the
first arg, just like it always has.

BUG=none
BRANCH=ToT, Samus
TEST=make runtests

In addtion to the new test included with this CL, I manually ran

  lddtree --copy-to-tree=blah --generate-wrappers /usr/bin/futility
  ./blah/usr/bin/futility dump_fmap -h tests/futility/data/bios_peppy_mp.bin

Before this CL, the wrapper didn't work because the binary was
being invoked as futility.elf, which was rejected. After this CL,
the wrapper works fine.

Change-Id: Iafdaff6e07ed294a7d29e4cff599ace0a3089229
Signed-off-by: Bill Richardson <wfrichar@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/223386
Reviewed-by: Hung-Te Lin <hungte@chromium.org>