AGESA binary PI: Add CONST to Azalia fields in callback structs
The Azalia table is a lookup. It is hard to imagine that it should
not be CONST.  The compiler does not complain when the Azalia
related fields in the structs passed into the AGESA OEM callout
are set CONST.  If the compiler does not complain, then the
calling function does not modify the Azalia lookup table.
Therefore, there is no issue with setting the Azalia verb table
pointer fields as CONST. All this does is provide more detail to the
compiler so that it can flag errors at compile-time rather than
runtime.

Change-Id: I269c137f8644e97e095e1e39df1a255223cf07b0
Signed-off-by: Bruce Griffith <Bruce.Griffith@se-eng.com>