vboot2: Split NV storage field defintions into their own header
And include it in vb2_api.h if VB20 internals are needed.

This allows coreboot to get at the fields more cleanly for now, rather
than duplicating the field definitions.  In the long run, we should
have APIs for this rather than having coreboot peek at the bits
directly.

BUG=none
BRANCH=none
TEST=emerge-veyron_pinky coreboot && make -j runtests

Change-Id: Ic308c3470773b91191bf682ff1b3cfce8864d26a
Signed-off-by: Randall Spangler <rspangler@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/240285
Reviewed-by: Bill Richardson <wfrichar@chromium.org>