Convert microcode to binary
Intel's microcode is a binary that has been converted to
an array of 32bit values. Instead of converting that back
to binary on every compile, just store the microcode as
a binary in the first place, and add that binary to CBFS.

This patch adds a script that takes all current Intel microcode.h
files and produces microcode.bin files. In addition, this patch
also adds all the microcode.bin files produced by the script.

Change-Id: Ia2712b50b49685f3eb781c0c68168ea1914350f8
Signed-off-by: Stefan Reinauer <stefan.reinauer@coreboot.org>