gma: Limit HDMI pixel rate
Given that current Intel chips only support 4K@30Hz over HDMI, it's
actually no surprise that older chips had even lower limits. For a
24bpp mode, it's 225MHz for Nehalem, Sandy and Ivy Bridge and 300MHz
for everything newer.

To keep the code most simple we keep the display's native resolution
and just limit the clock if we encouter an unsupported mode.

TEST=Booted T420 with a DELL U3415W attached over DP-to-HDMI and
     DVI-to-HDMI adapters. The display came up with 3440x1440@42Hz
     (instead of a garbled picture at 1720x1440@60Hz).

Change-Id: I5386386c648db267a49ed1f62750d2a5fdb2d934
Signed-off-by: Nico Huber <nico.h@gmx.de>
Reviewed-on: https://review.coreboot.org/17499
Reviewed-by: Arthur Heymans <arthur@aheymans.xyz>
Reviewed-by: Paul Menzel <paulepanter@users.sourceforge.net>