Add muen scheduling info as timer source
This is a very low precision clock. We just take the start of the cur-
rent scheduling minor frame as the minimum, yet elapsed moment and the
end of the frame as the maximum, yet reached moment in time. As a
result, every timer based delay will wait at least for the next minor
frame.

For the build, `musinfo.ads` and `muschedinfo.ads` from the Muen
project are required. One should point the `muen-common-path` variable
to the `common` subdir of the Muen source distribution.

Change-Id: Ib139749214bf8d2ca293e31327b156de5198b65d
Signed-off-by: Nico Huber <nico.huber@secunet.com>
Reviewed-on: https://review.coreboot.org/18357
Tested-by: Nico Huber <nico.h@gmx.de>
Reviewed-by: Adrian-Ken Rueegsegger <ken@codelabs.ch>